Move stringify into its own file.
